requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
freezegun==1.2.2
graphene==3.2.1
graphql-core==3.2.3
graphql-relay==3.2.0
Pillow==9.4.0
promise==2.3
pycodestyle==2.10.0
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
